impl decorator pattern
