+++ src/xpring-client.ts
@@ -1,40 +1,15 @@
-import {
-  AccountInfo,
-  GetAccountInfoRequest,
-  GetFeeRequest,
-  GetLatestValidatedLedgerSequenceRequest,
-  Payment,
-  Signer,
-  SubmitSignedTransactionRequest,
-  Transaction,
-  Utils,
-  Wallet,
-  XRPAmount
-} from "xpring-common-js";
-import { NetworkClient } from "./network-client";
-import GRPCNetworkClient from "./grpc-network-client";
+import { Wallet } from "xpring-common-js";
+import { XpringClientDecorator } from "./xpring-client-decorator";
+import DefaultXpringClient from "./default-xpring-client";
 
 /* global BigInt */
 
 /**
- * Error messages from XpringClient.
- */
-export class XpringClientErrorMessages {
-  public static readonly malformedResponse = "Malformed Response.";
-  public static readonly signingFailure = "Unable to sign the transaction";
-  /* eslint-disable  @typescript-eslint/indent */
-  public static readonly xAddressRequired =
-    "Please use the X-Address format. See: https://xrpaddress.info/.";
-  /* eslint-enable  @typescript-eslint/indent */
-}
-
-/** A margin to pad the current ledger sequence with when submitting transactions. */
-const ledgerSequenceMargin = 10;
-
-/**
  * XpringClient is a client which interacts with the Xpring platform.
  */
-class XpringClient {
+class XpringClient implements XpringClientDecorator {
+  private readonly decoratedClient: XpringClientDecorator;
+
   /**
    * Create a new XpringClient.
    *
@@ -42,19 +17,11 @@ class XpringClient {
    *
    * @param grpcURL The URL of the gRPC instance to connect to.
    */
-  public static xpringClientWithEndpoint(grpcURL: string): XpringClient {
-    const grpcClient = new GRPCNetworkClient(grpcURL);
-    return new XpringClient(grpcClient);
+  public constructor(grpcURL: string) {
+    this.decoratedClient = DefaultXpringClient.defaultXpringClientWithEndpoint(
+      grpcURL
+    );
   }
-
-  /**
-   * Create a new XpringClient with a custom network client implementation.
-   *
-   * In general, clients should prefer to call `xpringClientWithEndpoint`. This constructor is provided to improve testability of this class.
-   *
-   * @param networkClient A network client which will manage remote RPCs to Rippled.
-   */
-  public constructor(private readonly networkClient: NetworkClient) {}
 
   /**
    * Retrieve the balance for the given address.
@@ -63,67 +30,8 @@ class XpringClient {
    * @returns A `BigInt` representing the number of drops of XRP in the account.
    */
   public async getBalance(address: string): Promise<BigInt> {
-    if (!Utils.isValidXAddress(address)) {
-      return Promise.reject(
-        new Error(XpringClientErrorMessages.xAddressRequired)
-      );
-    }
-
-    return this.getAccountInfo(address).then(async accountInfo => {
-      const balance = accountInfo.getBalance();
-      if (balance === undefined) {
-        return Promise.reject(
-          new Error(XpringClientErrorMessages.malformedResponse)
-        );
-      }
-
-      return BigInt(balance.getDrops());
-    });
+    return await this.decoratedClient.getBalance(address);
   }
-
-  /* eslint-disable no-dupe-class-members */
-
-  /**
-   * Send the given amount of XRP from the source wallet to the destination address.
-   *
-   * @param drops A `BigInt` representing the number of drops to send.
-   * @param destination A destination address to send the drops to.
-   * @param sender The wallet that XRP will be sent from and which will sign the request.
-   * @returns A promise which resolves to a string representing the hash of the submitted transaction.
-   */
-  public async send(
-    amount: BigInt,
-    destination: string,
-    sender: Wallet
-  ): Promise<string>;
-
-  /**
-   * Send the given amount of XRP from the source wallet to the destination address.
-   *
-   * @param drops A numeric string representing the number of drops to send.
-   * @param destination A destination address to send the drops to.
-   * @param sender The wallet that XRP will be sent from and which will sign the request.
-   * @returns A promise which resolves to a string representing the hash of the submitted transaction.
-   */
-  public async send(
-    amount: string,
-    destination: string,
-    sender: Wallet
-  ): Promise<string>;
-
-  /**
-   * Send the given amount of XRP from the source wallet to the destination address.
-   *
-   * @param drops A number representing the number of drops to send.
-   * @param destination A destination address to send the drops to.
-   * @param sender The wallet that XRP will be sent from and which will sign the request.
-   * @returns A promise which resolves to a string representing the hash of the submitted transaction.
-   */
-  public async send(
-    amount: number,
-    destination: string,
-    sender: Wallet
-  ): Promise<string>;
 
   /**
    * Send the given amount of XRP from the source wallet to the destination address.
@@ -138,127 +46,7 @@ class XpringClient {
     destination: string,
     sender: Wallet
   ): Promise<string> {
-    if (!Utils.isValidXAddress(destination)) {
-      return Promise.reject(
-        new Error(XpringClientErrorMessages.xAddressRequired)
-      );
-    }
-
-    const normalizedAmount = this.toBigInt(amount);
-
-    return this.getFee().then(async fee => {
-      return this.getAccountInfo(sender.getAddress()).then(
-        async accountInfo => {
-          return this.getLastValidatedLedgerSequence().then(
-            async ledgerSequence => {
-              if (accountInfo.getSequence() === undefined) {
-                return Promise.reject(
-                  new Error(XpringClientErrorMessages.malformedResponse)
-                );
-              }
-
-              const xrpAmount = new XRPAmount();
-              xrpAmount.setDrops(normalizedAmount.toString());
-
-              const payment = new Payment();
-              payment.setXrpAmount(xrpAmount);
-              payment.setDestination(destination);
-
-              const transaction = new Transaction();
-              transaction.setAccount(sender.getAddress());
-              transaction.setFee(fee);
-              transaction.setSequence(accountInfo.getSequence());
-              transaction.setPayment(payment);
-              transaction.setLastLedgerSequence(
-                ledgerSequence + ledgerSequenceMargin
-              );
-              transaction.setSigningPublicKeyHex(sender.getPublicKey());
-
-              var signedTransaction;
-              try {
-                signedTransaction = Signer.signTransaction(transaction, sender);
-              } catch (signingError) {
-                const signingErrorMessage =
-                  XpringClientErrorMessages.signingFailure +
-                  ". " +
-                  signingError.message;
-                return Promise.reject(new Error(signingErrorMessage));
-              }
-              if (signedTransaction == undefined) {
-                return Promise.reject(
-                  new Error(XpringClientErrorMessages.signingFailure)
-                );
-              }
-
-              const submitSignedTransactionRequest = new SubmitSignedTransactionRequest();
-              submitSignedTransactionRequest.setSignedTransaction(
-                signedTransaction
-              );
-
-              return this.networkClient
-                .submitSignedTransaction(submitSignedTransactionRequest)
-                .then(async response => {
-                  const transactionBlob = response.getTransactionBlob();
-                  const transactionHash = Utils.transactionBlobToTransactionHash(
-                    transactionBlob
-                  );
-                  if (!transactionHash) {
-                    return Promise.reject(
-                      new Error(XpringClientErrorMessages.malformedResponse)
-                    );
-                  }
-                  return Promise.resolve(transactionHash);
-                });
-            }
-          );
-        }
-      );
-    });
-  }
-
-  /* eslint-enable no-dupe-class-members */
-
-  private async getLastValidatedLedgerSequence(): Promise<number> {
-    const getLatestValidatedLedgerSequenceRequest = new GetLatestValidatedLedgerSequenceRequest();
-    const ledgerSequence = await this.networkClient.getLatestValidatedLedgerSequence(
-      getLatestValidatedLedgerSequenceRequest
-    );
-    return ledgerSequence.getIndex();
-  }
-
-  private async getAccountInfo(address: string): Promise<AccountInfo> {
-    const getAccountInfoRequest = new GetAccountInfoRequest();
-    getAccountInfoRequest.setAddress(address);
-    return this.networkClient.getAccountInfo(getAccountInfoRequest);
-  }
-
-  private async getFee(): Promise<XRPAmount> {
-    const getFeeRequest = new GetFeeRequest();
-
-    return this.networkClient.getFee(getFeeRequest).then(async fee => {
-      const feeAmount = fee.getAmount();
-      if (feeAmount == undefined) {
-        return Promise.reject(
-          new Error(XpringClientErrorMessages.malformedResponse)
-        );
-      }
-      return feeAmount;
-    });
-  }
-
-  /**
-   * Convert a polymorphic numeric value into a BigInt.
-   *
-   * @param value The value to convert.
-   * @returns A BigInt representing the input value.
-   */
-  private toBigInt(value: string | number | BigInt): BigInt {
-    if (typeof value == "string" || typeof value == "number") {
-      return BigInt(value);
-    } else {
-      // Value is already a BigInt.
-      return value;
-    }
+    return await this.decoratedClient.send(amount, destination, sender);
   }
 }
