+++ test/default-xpring-client-test.ts
@@ -6,7 +6,9 @@ import { Utils, Wallet, WalletGenerationResult } from "xpring-common-js";
 
 import chai from "chai";
 import chaiString from "chai-string";
-import XpringClient, { XpringClientErrorMessages } from "../src/xpring-client";
+import DefaultXpringClient, {
+  XpringClientErrorMessages
+} from "../src/default-xpring-client";
 import {
   FakeNetworkClient,
   FakeNetworkClientResponses
@@ -22,10 +24,10 @@ chai.use(chaiString);
 
 const testAddress = "X76YZJgkFzdSLZQTa7UzVSs34tFgyV2P16S3bvC8AWpmwdH";
 
-describe("Xpring Client", function(): void {
+describe("Default Xpring Client", function(): void {
   it("Get Account Balance - successful response", async function() {
-    // GIVEN a XpringClient.
-    const xpringClient = new XpringClient(fakeSucceedingNetworkClient);
+    // GIVEN a DefaultXpringClient.
+    const xpringClient = new DefaultXpringClient(fakeSucceedingNetworkClient);
 
     // WHEN the balance for an account is requested.
     const balance = await xpringClient.getBalance(testAddress);
@@ -36,7 +38,7 @@ describe("Xpring Client", function(): void {
 
   it("Get Account Balance - classic address", function(done) {
     // GIVEN a XpringClient and a classic address
-    const xpringClient = new XpringClient(fakeSucceedingNetworkClient);
+    const xpringClient = new DefaultXpringClient(fakeSucceedingNetworkClient);
     const classicAddress = "rsegqrgSP8XmhCYwL9enkZ9BNDNawfPZnn";
 
     // WHEN the balance for an account is requested THEN an error to use X-Addresses is thrown.
@@ -49,7 +51,7 @@ describe("Xpring Client", function(): void {
 
   it("Get Account Balance - error", function(done) {
     // GIVEN a XpringClient which wraps an erroring network client.
-    const xpringClient = new XpringClient(fakeErroringNetworkClient);
+    const xpringClient = new DefaultXpringClient(fakeErroringNetworkClient);
 
     // WHEN a balance is requested THEN an error is propagated.
     xpringClient.getBalance(testAddress).catch(error => {
@@ -67,7 +69,7 @@ describe("Xpring Client", function(): void {
       accountInfoResponse
     );
     const fakeNetworkClient = new FakeNetworkClient(fakeNetworkClientResponses);
-    const xpringClient = new XpringClient(fakeNetworkClient);
+    const xpringClient = new DefaultXpringClient(fakeNetworkClient);
 
     // WHEN a balance is requested THEN an error is propagated.
     xpringClient.getBalance(testAddress).catch(error => {
@@ -79,7 +81,7 @@ describe("Xpring Client", function(): void {
 
   it("Send XRP Transaction - success with BigInt", async function() {
     // GIVEN a XpringClient, a wallet, and a BigInt denomonated amount.
-    const xpringClient = new XpringClient(fakeSucceedingNetworkClient);
+    const xpringClient = new DefaultXpringClient(fakeSucceedingNetworkClient);
     const wallet = (Wallet.generateRandomWallet() as WalletGenerationResult)
       .wallet;
     const destinationAddress =
@@ -105,7 +107,7 @@ describe("Xpring Client", function(): void {
 
   it("Send XRP Transaction - success with number", async function() {
     // GIVEN a XpringClient, a wallet, and a number denominated amount.
-    const xpringClient = new XpringClient(fakeSucceedingNetworkClient);
+    const xpringClient = new DefaultXpringClient(fakeSucceedingNetworkClient);
     const wallet = (Wallet.generateRandomWallet() as WalletGenerationResult)
       .wallet;
     const destinationAddress =
@@ -131,7 +133,7 @@ describe("Xpring Client", function(): void {
 
   it("Send XRP Transaction - success with string", async function() {
     // GIVEN a XpringClient, a wallet, and a numeric string denominated amount.
-    const xpringClient = new XpringClient(fakeSucceedingNetworkClient);
+    const xpringClient = new DefaultXpringClient(fakeSucceedingNetworkClient);
     const wallet = (Wallet.generateRandomWallet() as WalletGenerationResult)
       .wallet;
     const destinationAddress =
@@ -157,7 +159,7 @@ describe("Xpring Client", function(): void {
 
   it("Send XRP Transaction - failure with invalid string", function(done) {
     // GIVEN a XpringClient, a wallet and an amount that is invalid.
-    const xpringClient = new XpringClient(fakeSucceedingNetworkClient);
+    const xpringClient = new DefaultXpringClient(fakeSucceedingNetworkClient);
     const wallet = (Wallet.generateRandomWallet() as WalletGenerationResult)
       .wallet;
     const destinationAddress =
@@ -179,7 +181,7 @@ describe("Xpring Client", function(): void {
       FakeNetworkClientResponses.defaultSubmitSignedTransactionResponse()
     );
     const feeFailingNetworkClient = new FakeNetworkClient(feeFailureResponses);
-    const xpringClient = new XpringClient(feeFailingNetworkClient);
+    const xpringClient = new DefaultXpringClient(feeFailingNetworkClient);
     const wallet = (Wallet.generateRandomWallet() as WalletGenerationResult)
       .wallet;
     const destinationAddress =
@@ -199,7 +201,7 @@ describe("Xpring Client", function(): void {
 
   it("Send XRP Transaction - failure with classic address", function(done) {
     // GIVEN a XpringClient, a wallet, and a classic address as the destination.
-    const xpringClient = new XpringClient(fakeSucceedingNetworkClient);
+    const xpringClient = new DefaultXpringClient(fakeSucceedingNetworkClient);
     const wallet = (Wallet.generateRandomWallet() as WalletGenerationResult)
       .wallet;
     const destinationAddress = "rsegqrgSP8XmhCYwL9enkZ9BNDNawfPZnn";
@@ -221,7 +223,7 @@ describe("Xpring Client", function(): void {
       FakeNetworkClientResponses.defaultSubmitSignedTransactionResponse()
     );
     const feeFailingNetworkClient = new FakeNetworkClient(feeFailureResponses);
-    const xpringClient = new XpringClient(feeFailingNetworkClient);
+    const xpringClient = new DefaultXpringClient(feeFailingNetworkClient);
     const wallet = (Wallet.generateRandomWallet() as WalletGenerationResult)
       .wallet;
     const destinationAddress =
@@ -248,7 +250,7 @@ describe("Xpring Client", function(): void {
       FakeNetworkClientResponses.defaultError
     );
     const feeFailingNetworkClient = new FakeNetworkClient(feeFailureResponses);
-    const xpringClient = new XpringClient(feeFailingNetworkClient);
+    const xpringClient = new DefaultXpringClient(feeFailingNetworkClient);
     const wallet = (Wallet.generateRandomWallet() as WalletGenerationResult)
       .wallet;
     const destinationAddress =
@@ -274,7 +276,7 @@ describe("Xpring Client", function(): void {
       FakeNetworkClientResponses.defaultError
     );
     const feeFailingNetworkClient = new FakeNetworkClient(feeFailureResponses);
-    const xpringClient = new XpringClient(feeFailingNetworkClient);
+    const xpringClient = new DefaultXpringClient(feeFailingNetworkClient);
     const wallet = (Wallet.generateRandomWallet() as WalletGenerationResult)
       .wallet;
     const destinationAddress =
